Refactor relation-list route
