+++ packages/strapi-plugin-content-manager/controllers/ContentManager.js
@@ -342,56 +342,4 @@ module.exports = {
 
     ctx.body = pm.sanitize(unpublishedEntry, { action: ACTIONS.read });
   },
-
-  async findRelationList(ctx) {
-    const { model, targetField } = ctx.params;
-    const { _component, ...query } = ctx.request.query;
-
-    const contentManagerServices = strapi.plugins['content-manager'].services;
-
-    if (!targetField) {
-      return ctx.badRequest();
-    }
-
-    const modelDef = _component ? strapi.db.getModel(_component) : strapi.db.getModel(model);
-
-    if (!modelDef) {
-      return ctx.notFound('model.notFound');
-    }
-
-    const attr = modelDef.attributes[targetField];
-    if (!attr) {
-      return ctx.badRequest('targetField.invalid');
-    }
-
-    const target = strapi.db.getModelByAssoc(attr);
-
-    if (!target) {
-      return ctx.notFound('target.notFound');
-    }
-
-    const contentManagerService = contentManagerServices.contentmanager;
-
-    let entities = [];
-
-    if (_.has(ctx.request.query, '_q')) {
-      entities = await contentManagerService.search(target.uid, query);
-    } else {
-      entities = await contentManagerService.fetchAll(target.uid, query);
-    }
-
-    if (!entities) {
-      return ctx.notFound();
-    }
-
-    const modelConfig = _component
-      ? await contentManagerServices.components.getConfiguration(modelDef.uid)
-      : await contentManagerServices.contenttypes.getConfiguration(modelDef.uid);
-
-    const field = _.get(modelConfig, `metadatas.${targetField}.edit.mainField`, 'id');
-    const pickFields = [field, 'id', target.primaryKey, PUBLISHED_AT_ATTRIBUTE];
-    const sanitize = d => _.pick(d, pickFields);
-
-    ctx.body = _.isArray(entities) ? entities.map(sanitize) : sanitize(entities);
-  },
 };
